feat: format phone numbers with spaces every 3 digits for readability (#832)
Adds formatPhoneNumber helper in src/lib/phone.ts that detects the
international dial code and groups remaining digits in chunks of 3
(e.g. "+48793904950" → "+48 793 904 950"). Country/dial-code list is
extracted from phone-input.tsx into the shared module so both input
parsing and display formatting share one source of truth.

Applied at display sites: patient/employee/contact/company detail
pages, list tables, appointment preview popover, and appointment-form
patient picker. Form inputs and tel: hrefs intentionally keep the raw
stored value.
